Rename the items of  the reciprocating machines
